x86_64 kernel - Increase buffer cache and vnode resources, and more.
* Increase the maximum buffer cache from 200M to 400M.  Note that
  the buffer cache is backed by the VM page cache which is unlimited.

* Use size_t for kmalloc() tracking

* Allow 0 to be specified for kmalloc_raise_limit() which makes a
  kmalloc pool unlimited.

* Adjust the kern.maxvnodes autocalculation for both i386 and x86_64.
  i386 boxes with maximum memory will get a slightly lower vnode
  limit while x86_64 boxes will get a dramatically higher vnode limit.

* Remove kmalloc pool limits for vnodes, for HAMMER inodes, and
  for UFS inodes.  These pools track maxvnodes and do not require
  limits.

  This fixes occassional kmalloc assertions and allows the sysop to
  raise kern.maxvnodes on a running system.

+++ sys/sys/malloc.h
@@ -107,9 +107,9 @@
  */
 struct malloc_type {
 	struct malloc_type *ks_next;	/* next in list */
-	long 	ks_memuse[SMP_MAXCPU];	/* total memory held in bytes */
-	long	ks_loosememuse;		/* (inaccurate) aggregate memuse */
-	long	ks_limit;	/* most that are allowed to exist */
+	size_t 	ks_memuse[SMP_MAXCPU];	/* total memory held in bytes */
+	size_t	ks_loosememuse;		/* (inaccurate) aggregate memuse */
+	size_t	ks_limit;	/* most that are allowed to exist */
 	long	ks_size;	/* sizes of this thing that are allocated */
 	long	ks_inuse[SMP_MAXCPU]; /* # of allocs currently in use */
 	__int64_t ks_calls;	/* total packets of this type ever allocated */
